Implement volume control on iOS and WatchOS
Added a volume property to the SoundManager class; then I created an
AppStorage property 'promptVolume' to ContentView w/ corresponding
binding to SettignsView.  When the user changes this property via the
slider in SettingsView, it updates the volume property in SoundManager.

I have added an onAppear to ContentView to load the promptVolume
property into SoundManger.instance.volume when the app loads.

Did this for both iOS and WatchOS versions of the app and they work.