Further work on updating the UI for the app
I have broken down the main visual component (the countdown gauge)
into multiple subviews.  Part of this includes moving the mechanisms
that control the property manipulation from the ContentView into the
TimerCompositeView, which then feeds the data into the MainArcView
and the AnalogView.

I am now working on making the Arc more customizable so that when the
stretchPhase changes and the updateAngle (fka updateEndAngle) triggers
a redraw of the view, it works like I expect it to.